test: refactor - move arrow flight server mocks to util/mocks.py package
